refactor: Split texture_converter.py and model_converter.py into subpackages (#1103)
Pure decomposition, no logic changes. Public __all__ exports remain identical.

texture_converter/ (was 57K monolithic):
- atlas.py: Texture atlas detection and extraction, metadata parsing
- validation.py: validate_texture, validate_textures_batch
- path_mapper.py: convert_java_texture_path, Java→Bedrock mapping
- fallback.py: Fallback texture generation
- jar_extractor.py: JAR texture extraction
- conversion.py: Core conversion logic, batch processing

model_converter/ (was 32K monolithic):
- geometry.py: convert_single_model, generate_model_structure
- blockstate.py: parse_blockstate, convert_blockstate
- item_model.py: Item-specific model handling
- inheritance.py: Parent model resolution
- jar_extractor.py: JAR model extraction

Legacy .py files updated to re-export from packages (backwards compatible).
Updated pyproject.toml with new directory ignores. All 19 tests pass.
