+++ ai-engine/pyproject.toml
@@ -92,6 +92,8 @@ ignore = [
 "agents/java_analyzer/" = ["N806"]  # Variable in function should be lowercase (RuntimeVisibleAnnotations) - split into submodules
 "tools/search_tool.py" = ["N806"]  # Variable in function should be lowercase (FallbackToolClass)
 "agents/audio_converter.py" = ["F401", "F822"]  # Optional converter_utils import, __all__ references private functions
+"agents/model_converter/" = ["F401", "C901", "F822"]  # Package re-exports, complex module
+"agents/texture_converter/" = ["F401", "C901", "F822"]  # Package re-exports, complex module
 "agents/model_converter.py" = ["F401", "C901", "F822"]  # Optional converter_utils import, __all__ references private functions
 "agents/texture_converter.py" = ["F401", "C901", "F822"]  # Optional converter_utils import, __all__ references private functions
 "agents/packaging_agent.py" = ["C901"]  # Legacy complexity
